A Reworked Version of Themes
Light and dark theming is reworked and abstracted to a separate module. The new module can detect when the system theme changes and update the UI. Additionally, when this happens it is also capable of notifying listeners of the update so the UI can react. A theme switcher which accurately displays the applied theme name is added (but not styled).