Making Space for Telemetry, Using Event Listener for Theme Change
There's a telemetry lib now which does logging and could send telemetry in the future. Use this instead of logging.

Using an event listener for system theme changes so we can be expedient about applying the system preferred them when the user does not declare a preference in the app (the system is their preference).